% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitExperimentalPlan.R
\name{InitExperimentalPlan}
\alias{InitExperimentalPlan}
\title{Initialize an experimental design object}
\usage{
InitExperimentalPlan(n)
}
\arguments{
\item{n}{Number of participant you want in your experiment.}
}
\value{
Returns an experimental design as form of a data frame.
}
\description{
Initialize an experimental design object
}
\examples{
InitExperimentalPlan(80)
}

