% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{is.ExperimentalPlan}
\alias{is.ExperimentalPlan}
\title{Test if the object is an ExperimentalPlan.}
\usage{
is.ExperimentalPlan(x)
}
\arguments{
\item{x}{An object}
}
\value{
\code{TRUE} if the object inherits from the \code{ExperimentalPlan} class.
}
\description{
Test if the object is an ExperimentalPlan.
}

