% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomizeOrder.R
\name{RandomizeOrder}
\alias{RandomizeOrder}
\title{Randomize order of your participant}
\usage{
RandomizeOrder(object)
}
\arguments{
\item{object}{An experimental plan object.}
}
\value{
A randomly ordered experimental plan.
}
\description{
Randomize order of your participant
}
\examples{
Condition <-
  c("Eperimental", "Control")

InitExperimentalPlan(80) \%>\%
 AddFactor(Condition) \%>\%
 RandomizeOrder()
}

