% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveExperimentalPlan.R
\name{SaveExperimentalPlan}
\alias{SaveExperimentalPlan}
\title{Saves experimental plan as a file.}
\usage{
SaveExperimentalPlan(data, name, type = "xlsx", path = NA)
}
\arguments{
\item{data}{An experimental plan object.}

\item{name}{What should the created file be named.}

\item{type}{What kind of file should be created (csv, xlsx, or e-prime inline).}

\item{path}{(Optional) A path to where the created file should be saved.}
}
\description{
Saves experimental plan as a file.
}
\examples{
Condition <-
  c("Eperimental", "Control")

InitExperimentalPlan(80) \%>\%
 AddFactor(Condition) \%>\%
 RandomizeOrder() \%>\%
 SavecExperimentalPlan("myplan.xlsx")
}

