% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddFactor.R
\name{AddFactor}
\alias{AddFactor}
\title{Adds a factor to an experimental plan object}
\usage{
AddFactor(object, factor)
}
\arguments{
\item{object}{An experimental plan object.}

\item{factor}{Factor you wish to add to you experimental plan.
Factor must be specified as a vector object (see examples section).}
}
\value{
An experimental plan object.
}
\description{
Adds a factor to an experimental plan object
}
\examples{
Condition <-
  c("Eperimental", "Control")

InitExperimentalPlan(80) \%>\%
 AddFactor(Condition)
}

